Extract the hook to manage form buttons
