+++ site/pages/token-approvals.js
@@ -7,12 +7,12 @@ import debounce from 'lodash.debounce'
 import vesperMetadata from 'vesper-metadata'
 import useTranslation from 'next-translate/useTranslation'
 
+import { fromUnit, toUnit } from '../utils'
+import { useFormButton } from '../hooks/useFormButton'
 import Button from '../components/Button'
 import Input from '../components/Input'
 import Layout from '../components/Layout'
 import PureContext from '../components/context/Pure'
-
-import { fromUnit, toUnit } from '../utils'
 
 const extraTokens = [].concat(
   // Include all Vesper pool tokens and the VSP token.
@@ -200,43 +200,6 @@ const useFeedback = function () {
   return [feedback, setFeedback]
 }
 
-const useFormButton = function (disabled, setFeedback, onClick) {
-  const { t } = useTranslation('common')
-  const { active } = useWeb3React()
-
-  const [inProgress, setInProgress] = useState(false)
-
-  const handleClick = function () {
-    setInProgress(true)
-    setFeedback('info', t('approval-in-progress'))
-    onClick()
-      .then(function () {
-        setFeedback('success', t('approval-succeeded'))
-      })
-      .catch(function (err) {
-        setFeedback('error', err.message)
-      })
-      .finally(function () {
-        setInProgress(false)
-      })
-  }
-
-  useEffect(
-    function () {
-      if (active) {
-        return
-      }
-      setInProgress(false)
-    },
-    [active]
-  )
-
-  return {
-    disabled: disabled || inProgress,
-    onClick: handleClick
-  }
-}
-
 const TokenApprovalsForm = function () {
   const { t } = useTranslation('common')
   const { account } = useWeb3React()
@@ -259,22 +222,45 @@ const TokenApprovalsForm = function () {
 
   const [feedback, setFeedback] = useFeedback()
 
+  // Depending on the progress state of the approval operation, set the feedback
+  // color and message.
+  const onProgress = function (err, state) {
+    if (err) {
+      setFeedback('error', err.message)
+      return
+    }
+
+    const messages = {
+      info: t('approval-in-progress'),
+      success: t('approval-succeeded')
+    }
+    setFeedback(state, messages[state])
+  }
+
+  // Set the approval buttons behavior, linking to the contract calls and
+  // progress feedback.
   const approveDisabled = !allowance
-  const approveButton = useFormButton(approveDisabled, setFeedback, () =>
-    tokenApprovals.approve(
-      token.address,
-      spender.address,
-      toUnit(allowance, token.decimals)
-    )
+  const approveButton = useFormButton(
+    approveDisabled,
+    () =>
+      tokenApprovals.approve(
+        token.address,
+        spender.address,
+        toUnit(allowance, token.decimals)
+      ),
+    onProgress
   )
-  const infiniteButton = useFormButton(approveDisabled, setFeedback, () =>
-    tokenApprovals
-      .approveInfinite(token.address, spender.address)
-      .then(() =>
-        tokenApprovals
-          .allowance(token.address, account, spender.address)
-          .then(setAllowance)
-      )
+  const infiniteButton = useFormButton(
+    approveDisabled,
+    () =>
+      tokenApprovals
+        .approveInfinite(token.address, spender.address)
+        .then(() =>
+          tokenApprovals
+            .allowance(token.address, account, spender.address)
+            .then(setAllowance)
+        ),
+    onProgress
   )
 
   return (
